Don't unwrap polyt for type_filter
Summary:
Currently when a PolyT flows through a predicate, the PolyT is implicitly instantiated and unwrapped. This results in erroneous behavior in examples like the below:
```

  type S = <W>(W) => W;

  let s: ?S = null;
  if (s != null) {
      const h = s;
      h(42);
  }
```
Currently in LTI mode, we'll get an error when we pass `42` into `h`, because `h` expects `empty`. This arises from `s`'s type being implicitly instantiated when it flows into a PredicateT, and the resulting tvar for `W` being resolved to empty.

It isn't really necessary though to instantiate a polyT before filtering it, though, and this diff removes that behavior by passing PolyTs directly into `predicate` and extending `type_filter` to support them.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D41452131

fbshipit-source-id: 71b29fcf379f5ce64787afe853caf6944402a3bb