[lti][autocomplete] Unbreak upper-bound-based autocompletion in LTI mode
Summary:
In the old inference system, we use an upper-bound based mechanism to find the type of some placeholder expression. We achieve that by creating an empty tvar, and let it interact with other stuff to accumulate upper bounds. Then during autocompletion, we can inspect the upper bound to produce completion results.

This mechanism broke down in LTI, when we seal tvars at expression boundaries. To fix this, I will query the hint available at the location instead. The hint is fully resolved, and it will directly be the type of the placeholder expression. Later when we are querying the upper bound, the fully resolved type will be considered as the only upper bound.

Reviewed By: panagosg7

Differential Revision: D41626700

fbshipit-source-id: 7ec38d301351aafea15af78c976705727ca2bdf3