move multi-spec logic to type_hint.ml
Summary:
For most constraints generated during hint decomposition, we try to explore all branches
of union (or equivalent) types for potential hints. This feature, known as "multi-spec", was
implemented in D40170579 (https://github.com/facebook/flow/commit/d4bec5b4d3aee3eab462fc51c7d822403ad14629). Each time a constraint of the form `UnionT ~> _` (for eg.) was
encountered under hint-computation, a special kind of speculation mode was used that
allowed multiple successful branches.

The check of *when* to perform this logic was based on a global flag (`run_in_hint_decomp`).
This has an unwanted side-effect that any constraint of this form encountered during
the entire hint computation, including constraints unrelated to the specific decomposition,
would follow this new semantics.

To avoid this situation, this diff moves this logic away from the constraint solver (Flow_js
and Subtyping_kit) and into Type_hint. To do so it needs to determine when a type is a
union, maybe type, etc. and break it up into its parts. It does so, by extending the mechanism
of `PreprocessingKitT`, also used during intersection speculation. The new variant
`ConcretizeHintT` introduced here merely propagates all concrete parts of the resolved
type to a target tvar.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D41675904

fbshipit-source-id: a1fa5788938128fce7cdffdf6086f127c370ccd8