Allow type_filter to recurse into resolved OpenTs in unions
Summary:
Flow currently operates under the assumption that OpenTs aren't relevant when performing type filtering in refinements--either OpenTs won't have appeared in the first place because we're filtering an annotation, or the we would look through the bounds of the OpenT before passing anything to type_filter. LTI breaks this assumption, because now UnionTs don't only appear as annotations--we can resolve a variable to a union of OpenTs (which are themselves fully resolved). The simplest way to do that is `const x = c ? a : b`--the type of `x` will be a UnionT with the types of `a` and `b` as its members.

This means that currently we don't correctly filter types under LTI, leading to new errors--for example, in the test case this diff adds, we originally would report that `undefined` doesnt have a `filter` property, because the `??` operator can't propertly filter `thumbnails`. This diff fixes that by allowing Type_filter to recurse through OpenTs--but only through resolved ones.

Reviewed By: panagosg7

Differential Revision: D41724932

fbshipit-source-id: dd5b2c7108955b69281bd224ea112aa85be81df0