[lti][contextual] Give empty object function static to template function type used during hint decomposition
Summary:
During hint decomposition of FuncParam, we will produce a template function type, with the parameter we are interested in leaving as OpenT, and everything else is any. Then we flow its type to a function type we know to solve the OpenT. Notably, the function static is also left as any. Most of the times, this is harmless, but in the following case:

```
const _: React$AbstractComponent<Prop, mixed> = props => {...};
```

This will make `props` silently any! It happens because during `FunT ~> React$AbstractComponent`, we will try to get the component props default of `FunT`, which is any because it's in the function statics. Then the any type propagated to the solution of props and make everything any! As a result, we have quite a lot of coverage drop.

To avoid the issue, I give it an empty object function static instead. Most of the times we are contextually type an unannotated lambda as a simple expression, so this makes perfect sense.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision:
D41746582 (https://github.com/facebook/flow/commit/e06e16a3502008d0eb94ca58743ba91caedbf641)

------------------------------------------------------------------------
(from 1d9b298d591adadcc5d8cb3aa8bb3e6e876ab1d7)

fbshipit-source-id: 87d5f33ed376c284be37a9b14e30ab27ab578cfd