Change all providers barrier test to use haste modules
Summary:
We are going to remove the file "all providers" linked list, so rewrite this
test to use the haste equivalent.

This test was added in D39075057 (https://github.com/facebook/flow/commit/a635d2efffeb94d1bd0a24f17bce6bb28cbbf1b9) to ensure that we use a write barrier. The test
is still valid and useful for haste module providers.

Reviewed By: jbrown215

Differential Revision: D41539015

fbshipit-source-id: 4dc1738d78e997e435d3c539ca91e11cd6f192b0