Kill file modules
Summary:
Currently, file modules are only used to store the set of dependent files. This
diff moves that set to the file object itself.

I have two motivations for this work:

1. We stored file modules in the shared hash table, so using files instead uses
fewer hash table slots. This means we can keep the hash table small as repos
continue to grow.

2. Currently, we store resolved requires as serialized Modulename.t containing
the hash table keys for haste modules and files. I want to change this to use
addresses instead. Without this change, if we stored just an address to a file
module, it was difficult to get the file name because the file name was not
stored on the file module itself. Now, we can store the address of a file which
does have the file name.

This change also introduces a small complexity: "phantom" files, or files which
have no parse information, but need to exist because they represent a file
module.

We always store the dependent set on the .js file object. If there is only a
.js.flow file, we will create a phantom file object for the .js file which is
not parsed.

Similarly, if there are dependents but no providers, we will create a phantom
file object, similar to the phantom file modules we had before.

Because of this change, we might clear a file which still has dependents. When
this happens, we should not remove the file from the hash table. This diff just
never removes files from the hash table.

This behavior matches the previous behavior for file modules, which we also
never deleted. I will deal with this in follow-up work.

Reviewed By: mroch

Differential Revision: D41535037

fbshipit-source-id: ca4dd411dc987458436ffd93c442536c94b8997a