Use file parse information to get file module providers
Summary:
Path-based requires resolve to file modules. A file module is provided by either
a .js or a .js.flow file.

Currently, we track file module providers using the same technique as haste
modules: we have a "file module" object which points to a linked list of
provider files. The current provider is tracked using an entity which is updated
during commit modules.

This diff is the first step toward removing file modules. Unlike haste modules,
which can have arbitrary providers, a file module can only be provided by a .js
or .js.flow file.

Since there can only be two providers, we instead just look up each one in the
hash table. We can use the parse entities to determine which file is the
provider.

This diff only changes the way we find a provider. In follow-up diffs, we will
change how reverse dependencies are stored, then we will remove file modules.

Reviewed By: mroch

Differential Revision: D41539017

fbshipit-source-id: 058ff16f900f2ce11c9a4c71235798fbedbc6d9f