Allow exporting binary arith expressions without an annotation
Summary:
With all arithmetic operators now being overloaded, we can no longer assume that these expressions result in `number` in exports. D41619157 tried to solve this by requiring an annotation (as was already done for the existing overloaded operator, +), but this added a lot of errors, because it turns out people like to export constants like `export const SECONDS_IN_DAY = 60 * 60 * 24`. So instead, let's try to infer the types of these expressions.

Changelog:
[fix] Fixed binary arithmetic operations on bigints being exported/imported across modules as `any` or `number.`
[feature] An annotation is no longer required when exporting + expressions (e.g., `export default (3 + 3 : number)` can now just be `export default 3 + 3;`)

Reviewed By: samwgoldman

Differential Revision: D41756826

fbshipit-source-id: 0f70003a51a4281ea6796d21ecf0007a379ca536