Extract and simplify unary overload logic to Flow_js_utils
Summary:
Moving because this will be used in the next diff in annotation_inference.

This also simplifies a couple cases. Namely, the `any` cases are combined, and the final wildcard error case applies to all operators, not just update. The latter causes some `incompatible-use` errors to be changed to `unsafe-addition` (which will be renamed later).

Changelog: [internal]

Reviewed By: samwgoldman, gkz

Differential Revision: D41740170

fbshipit-source-id: 06519e11bbb5fc90dba8bf8cabd40745ce5bac7c