unify targs with bounds on MethodT NoMethodAction
Summary:
We use `MethodT NoMethodAction` constraints to access the type of the receiver of a
method call, typically in the case of hint computation. On polymorphic classes, this
constraint reuses the standard approach of instantiating the type parameters of the class.
(See code that changes in flow_js.ml around the call to `instantiate_poly`).
`instantiate_poly` instantiates the type arguments and flows them to the bounds of the
respective type parameter.

When the method action is `NoMethodAction`, this instantiation is bound to remain incomplete:
we are not going to get meaningful lower-bounds for the type arguments, since we won't
be performing the call as in other `MethodT` actions. This leaves the risk of ending up
with under-constrained tvars when resolving intermediate steps during hint decomposition.

So, instead, this diff unifies type arguments with the bounds when dealing with
`MethodT NoMethodAction`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D41814020

fbshipit-source-id: d425d8fe7509dc6c2f337528a0efdc2779b2894d