clear caches during reinit
Summary:
a reinit is similar to a recheck in that when its transaction commits, caches that pointed to the old data are now invalid. we cleared those caches in `Recheck.full` but not in `init_from_saved_state`.

previously, `init_from_saved_state` was only called during init, when we couldn't have populated the IDE caches with anything anyway; but now it's also called during reinit which is happening in parallel with IDE commands.

Reviewed By: samwgoldman

Differential Revision: D41806868

fbshipit-source-id: a810c3d36a8ebe4ffdd2960da24080aa91cb3301