fix string literal completion with existing quotes
Summary:
there's a config setting to decide whether to prefer single quotes. this was being used when printing string literal completions. if you typed `foo["|`, it would suggest `'bar'`. VSCode would filter this result out because it doesn't match the `"`.

vice versa is also true if single_quotes=false and you complete `foo['|`

the fix is, if the token starts with a quote, generate string literals with that quote.

Changelog: [fix] Fix autocomplete of string literals in bracket syntax, after a quote has already been typed

Reviewed By: darichey

Differential Revision: D41831205

fbshipit-source-id: 4e255c24d90eb6981f2e786e60da89989bf16ca8