add an LSP init flag for strict completion ordering
Summary:
vscode ignores the sort order that the server sends and reranks the results itself.

our vscode extension can mess with `filterText` such that vscode's fuzzy score is always the same, preventing it from reranking. it then breaks the tie using the server's `sortText`.

this diff adds an `experimental.strictCompletionOrder` server capability so that the server can tell the extension that it wants this behavior.

Changelog: [internal]

Reviewed By: darichey

Differential Revision: D41801836

fbshipit-source-id: a8cb3fe0e6bd25529882f0b2747ee5e2d251c385