extension and server must both opt into server controlled completion ordering
Summary:
we have a setting that forces vscode to obey the order of completion results as provided by the server, rather than reranking them itself.

the setting previously defaulted to `false`, but we need it to be `true` by default to roll it out. but if you're using an old version of flow where our ranking is bad, this is a regression.

D41801836 adds a server capability, `experimental.strictCompletionOrder`, so that the server can opt into this behavior.

this diff makes the extension only enable the hack if the server opts in, AND the setting is enabled -- which it now is by default. so, if you're on a new enough version of flow, you get server-ranked results by default but there's a killswitch; if you're on an old version, you don't get the server's shitty results.

Reviewed By: SamChou19815

Differential Revision: D41801837

fbshipit-source-id: aadb402ae577dee0210f73b1faf13b75ded4b5ef