Record reachable targs while substituting an ObjT to optimize any-propagation
Summary:
In this stack, I add a new field to ObjT that lets us more quickly find substituted targs inside an ObjT. This lets us avoid expensive deep traversals when doing any-propagation over an object.

This diff populates the field during substitution and uses it to do any propagation. When we enter an object, we set the reachable_targs to Some [] and add to the list when we see a non-generic substitution.  When we pop back up the stack to the ObjT again, we restore the old reachable_targs list and add our newly found targs, if applicable.

When we do any propagation we simply propagate to the reachable_targs instead of doing any deep traversals.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D41858132

fbshipit-source-id: 00ceec248948b54aedd31564270c33f0bcb75441