validate values of bool CLI flags
Summary:
bool cli flags currently treat anything that's not `0` or `false` as truthy, even `False` or `off`.

instead, only treat `0` and `false` as falsy, and `1` and `true` as truthy, and fail on everything else.

Changelog: [fix] Boolean CLI flags must now be `1` or `true` to be truthy. Before, anything other than `0` and `false` was considered truthy.

Reviewed By: jbrown215

Differential Revision: D41973790

fbshipit-source-id: 7ee693b1e5833376a99ebb0014d82a35b415c176