Fix code action service tests to read committed data
Summary:
Instead of rolling back every transaction, we need to commit the transaction
first, so we can use a State_reader to read the committed data back.

So, this diff changes the tests to commit transactions instead. After we commit,
we can run the function under test. After that, we run another transaction to
remove the file.

Reviewed By: mroch

Differential Revision: D42199510

fbshipit-source-id: c33c7b200eba62565e9584b7b3b2bdf194ab27aa