Further simplify prepare_add_checked_file
Summary:
This diff factors out a prepare function to set the parse data on a typed parse
object and calls this from each branch in this code.

Before, each branch would return a tuple (parse, dirty_modules) and we would
add the parse data when the branches joined back up.

In my opinion, this is nicer to read, even though every branch calls into the
same code. But it's pretty subjective, so willing to abandon this one.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D42254535

fbshipit-source-id: 6d1e1d1d34794691a00071e051d49c2c5609cc4b