Make updating haste info similar to updating resolved requires
Summary:
For both haste info and resolved requires, we only want to update the entity if
the underlying value is changed.

Before this diff, the function `prepare_write_new_haste_info_if_changed` was
used for both updates and newly created haste info. Furthermore, for the update
case, we needed to do two separate change tests, once to write the haste info
and again when updating the entity.

This diff splits out the create-new use case into a separate function, which is
also used by the update case if needed. Also, the update case also takes care of
updating the entity.

This is nicer overall, but also more similar to how we already did things for
resolved requires.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D42254530

fbshipit-source-id: b5433d8c088903decb1ec0d5d8445429fd5c2145