Make sharedmem allocation size calculation more regular
Summary:
Later in this stack, I am going to introduce a small abstraction around the
`prepare_*` functions here to make things more composable and nice to use.

To prepare for that, this diff makes the existing size calculation more regular
to fit that abstraction.

In particular, this diff stops accumulating the size in a `fold` style. instead,
we track all the sizes separtely and sum them once right before allocating.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D42254537

fbshipit-source-id: c4adc31c7a87c03d0786121a4d9c4b556d45d606