tail recursion in Name_resolver visit_program
Summary:
uses `List.rev_append` instead of the non-tail-recursive `@` in `Name_resolver`.

this is likely not an issue in practice, because `SMap.keys` is also not tail recursive, so if there were too many bindings for `@`, it'd have already crashed. but we're seeing stack overflow segfaults due to `@` (no idea which caller), so I'm auditing callers of `@` and would prefer to just avoid it.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D42344687

fbshipit-source-id: 37f4b4d01d0ddc4cd1ca246294b031a1e58ab434