Allow `mixed` or `any` as an annotation to a `catch` parameter (but not anything else)
Summary:
For

```
try {
  ...
} catch (e) {
  ...
}
```

Allow an optional annotation on `e` that is exactly `mixed` or `any`. The default is currently `any`, so in effect this would allow you to choose a type-safe `mixed`, and then refine the result. E.g.

```
try {
  ...
} catch (e: mixed) {
  if (e instanceof MyError) {
    ...
  } else {
    throw e;
  }
}
```

[TypeScript has the same behavior](https://www.typescriptlang.org/play?#code/C4JwngBA3gUDC+EDGBDYSAWEAUBTAXBAK4B2A1iQPYDuJAlNHPHKJLAsmpjgRAKIgQlEA3bwgA). If you choose something other than `mixed` or `any`, it gives you:

> Catch clause variable type annotation must be 'any' or 'unknown' if specified.

This is commonly requested. It was brought up in feedback meetings on Nov 2022, and the following support posts:
- https://fb.workplace.com/groups/flow/posts/3220196014695658/
- https://fb.workplace.com/groups/flow/posts/6053873331327898/
- https://fb.workplace.com/groups/flow/posts/4853249668056943/
- https://fb.workplace.com/groups/flow/posts/1235112129870733/
- https://fb.workplace.com/groups/flow/posts/9178694162179117/
- https://fb.workplace.com/groups/floweng/posts/6984961778219044/

Closes #3259
Closes #2470

TypeScript also allows aliasing, e.g.

```
type T = any;
try {
  ..
} catch (e: T) {
  ...
}
```

But we don’t need to do this, as it’s not that useful and would require more boilerplate code on our end (unless one of you has an idea for a quick solution).

Future work idea: Add `.flowconfig` option that changes the default from `any` to `mixed`. First we could codemod any non-annotated instances to `catch (e: any) {}`

Changelog: [feature] Allow `mixed` or `any` as an annotation to a `catch` parameter (but not anything else)

Reviewed By: SamChou19815

Differential Revision: D42158104

fbshipit-source-id: 6b306378599f8e5707c0cc728e9098ad5b74bd07