fix crash from unchecked deps during reinit
Summary:
we're seeing some `option is None` exceptions when `allow_reinit` is enabled (P582230678). this happens after a saved state reinit but before the recheck of local changes finishes (not sure the latter matters)

it fails when trying to do `get_type_sig parse` on a dependency's `parse` ent. first, it had to get past `ensure_checked_dependencies`, which says a dep is ok if it has a provider, typed parse, and leader, but doesn't care if it has a type_sig.

it is expected that we have a type sig in order to calculate the leader and sig_hash. this invariant was violated when reinit-ing from saved state: if the file has been checked before and has a leader, that leader was retained during reinit, but the type sig was erased. this allowed it to get through `ensure_checked_dependencies` incorrectly.

Changelog: [internal] allow_reinit is experimental and disabled by default

Reviewed By: samwgoldman

Differential Revision: D42169537

fbshipit-source-id: a3e61480fa46415edb4239f1dcfb9a19bdc69371