[normalizer] pass down ?id field
Summary:
This diff threads an optional `id` field down various functions in the normalizer.
This id corresponds to the latest OpenT or EvalT id seen. The id stops being
propagated as soon as we've seen a concrete type. Ie. it will descend down
UnionTs and IntersectionTs. The purpose of this change is to replace the
`remove_toplevel_tvar` implementation and perform the check that it currently
does on-the-fly.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D42024141

fbshipit-source-id: 8bd51ce76bef608ddb6d2ae6f6131249548fa415