[normalizer] detect trivially recursive types on the fly
Summary:
If the id of the currently inspected type is the same as the one last seen and
recorded in the `id` field of D42024141 (https://github.com/facebook/flow/commit/3b586c40524297316e9824a02fde6456333ee3ee), then this is a trivially recursive type.
The goal of this check is to convert the recursive portion of types like
`T = T | number` to `empty` and subsequently eliminate them from the
union they are part of.

Earlier this was happening in `remove_toplevel_tvar`. This code is deleted next.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D42024140

fbshipit-source-id: 8db5e4a6289806be2b9192b8b02338d4f0185e3c