Use polymorphic variants as phantom types for shared heap addrs
Summary:
Before this diff, we used abstract type as the phantom types for `'a addr`. This
works well enough when you only use addrs of a single type.

Sometimes we want to consider sets of addrs at once. For example, a file's parse
data might be typed, untyped, or a package json. For this, we use polymorphic
variants and the minimal subtyping they support.

This diff uses polymorphic variants for everything. My specific use case is to
write types like "haste module addr or file addr" as the type of a dependency.
In general, it's nice to be able to write the type of any union of addrs.

I have also defined some aliases to make this notationally more convenient. For
example:

```
[ `typed ] parse entity addr
```

instead of

```
[ `entity of [ `parse of [ `typed ] parse ] ] addr
```

Reviewed By: jbrown215

Differential Revision: D41560965

fbshipit-source-id: b00f5d6c37413ef3ab3aae36479616f58310d2b6