normalize "rest array" as tuple
Summary:
The type normalizer used to convert all inferred array literal types to `Array<T1 | T2 | ...>`
where `T1`, `T2`, etc. are the inferred types of the members of the array. This had as a
result in cases like
```
declare function foo<Args: $ReadOnlyArray<mixed>>(
    f: (...Args) => mixed,
  ): (...Args) => void;

declare var x: (string, number) => void;
foo(x)('a', 1);
```
the codemodded type for the instantiation of `Args` in the call to `foo` to be
`Array<string | number>` which causes a Flow error if annotated
```
foo<Array<string | number>>(x)('a', 1); // error: string incompat. with number, etc.
```
If instead we annotated with a tuple type this error would go away
```
foo<[string, number]>(x)('a', 1);
```
This is what this diff does.

Reviewed By: SamChou19815

Differential Revision: D42173801

fbshipit-source-id: a90c2d647d85f1e5664e61c13c7f513ffd419d26