fix fuzzy scores of 0
Summary:
fuzzy scores can be 0 and still be a match.

the vscode implementation returns `undefined` for no match, and `[0]` for a match with score 0. it tests `if (match)`, but `[0]` is truthy in JS.

likewise, we now return `None` for no match and `Some 0` for a match with score 0.

Changelog: [fix] Fix some missing autocomplete results (regression in 0.196.0)

Reviewed By: samwgoldman

Differential Revision: D42123744

fbshipit-source-id: 8793aed338c60d971189794c78ab98b299d41872