improve error recovery of type params
Summary:
when parsing a type param list, if we see an unexpected token, we consume it and create a bogus type param with the name `""`.

this causes problems downstream when we try to typecheck this ast. we could make all downstream analyses more robust, but it would be hard and messy.

this diff avoids creating invalid type params. instead, we just leave them out of the AST even though we consume the unexpected token.

this diff also adds a bit of special casing to bail out of the list when the unexpected token is something that commonly follows a param list, like a paren (like if you're adding params to an existing `function f<|()`) or class heritage (like `class C<| extends D ...`)

Changelog: [fix] Improve the parser's error recovery when in the middle of adding type parameters.

Reviewed By: samwgoldman

Differential Revision: D42247757

fbshipit-source-id: 32c92b10a0604e25a5517b626b3c4702ece0fad8