Only store mapped names in resolved module error case
Summary:
For resolved modules, the error case carries a string. However, this is usually
the same as the module ref which we already store. Why store it twice?

Well, I said "usually." If there is a module name mapper which fails to resolve,
then we store the mapped name.

Most of the time, though, the string is the same as the module reference.
Furthermore, we hit the error case reasonably often because it also encompasses
cases where a module ref resolved to a builtin module.

This diff changes the error case to be a `string option` instead, so we store
the mapped name only and never a duplicate of the module ref.

Reviewed By: mroch

Differential Revision: D42180155

fbshipit-source-id: 47da97917f8fca5231ff339a7ffc82b1e2377af3