Store phantom dependencies as addrs
Summary:
Instead of Modulename.Set.t, we can store the addresses of the haste/file
modules directly. We store the addrs in increasing order, so we can efficiently
reconstruct a map using `of_increasing_iterator_unchecked`.

For now, to minimize the changes here, we reconstruct the Modulename.Set.t when
we read these out. In the future, we can avoid this and use the addrs directly.

Reviewed By: mroch

Differential Revision: D42026419

fbshipit-source-id: d864894326ff0796fd0b018de7d1aa7700348551