Add module references to sharedmem and saved state
Summary:
The `resolved_modules` field of the `resolved_requires` record is currently a
mapping from module refs to resolved modules which are represented as
Modulename.t.

This means we serialize haste module names and file names for dependencies in
the heap for each dependency edge. On lookup, we need to take those strings and
look them up in the hash table.

In the following stack of diffs, I am going to change this, so we store
dependency edges as addresses directly.

As the first step, I am going to move the string keys to a separate object.
Later, we will store the resolved modules as a parallel array of addresses, and
recover the mapping by combining these two arrays.

For now, this is a regression since we store the module references twice: once
in the serialized resolved requires, and again for this new data. This will be
resolved later in the stack.

Reviewed By: mroch

Differential Revision: D42182951

fbshipit-source-id: 1c491dbfc6b7357266db82e40a144d84cc11ed2e