Separate resolved modules and phantom dependencies
Summary:
Instead of storing the entire `resolved_requires` record together, this diff
stores them separately.

This is prep for storing the phantom dependencies and resolved modules as
addresses instead of serialized blobs themselves, which will happen later in
this stack.

For now, this is likely a regression in heap size, since compression of the
combined serialized data would eliminate lots of duplication. Once we store
this data as addrs instead of strings, the difference won't matter.

Reviewed By: mroch

Differential Revision: D42014335

fbshipit-source-id: 4d47b08b37ef977080be4331c8fda140d91144fa