Stop tracking changes to resolved requires
Summary:
We used to use "resolved_requires_changed" to prevent skipping direct
dependents. Now that we no longer need that, we can also stop tracking this
information entirely.

Instead of hashing, this diff uses a simpler equality comparison when resolved
requires change. This is just to re-use the existing resolved requires instead
of allocating the same data again in shared memory.

There isn't a ton of value to this right now, but this cleanup enables work
which will be submitted in later diffs.

Reviewed By: mroch

Differential Revision: D42132291

fbshipit-source-id: 00f56d3b6a838331a439b0d08a4565ec665447f2