[codemod infra] Use projections of exported types for GraphQL fragments
Summary:
Given a GraphQL file `foo.graphql.js` exporting a type
```
export type Foo = {
  +f?: {
    +g: {
      +h: ?$ReadOnlyArray<{
        +i: {}
      }>
    };
  }
};
declare export function foo(): Foo;
```
and a file importing from `foo.graphql.js` and using the imports as
```
function m(x) {}
const x = foo()?.f?.g?.h?.[0]?.i;
if (x) {
  m(x);
}
```
a codemod that would try to annotate `m`'s `x` parameter, would hard-code the type `{}`.
In real world examples this type is typically much larger, leading to a bad codemodding
experience.

This diff changes this approach by trying to match the target type `{}` with a projection
of an existing type in `foo.graphql.js`. In this case the result would be
```
$NonMaybeType<Foo?.["f"]?.["g"]?.["h"]?.[0]?.["i"]>
```
It does so by visiting the type aliases defined in the AST of the GraphQL file and looking for
the def-loc of the type that we would otherwise codemod in. As it visits the AST, it
accumulates a `Ty.t` type that describes the currently visited type at each point in terms of
the alias name. When the target location is found, the accumulated type is returned and
replaces the inferred one.

Some features of the implementation:
* To simplify the logic that decides whether we use index access or optional index access,
once we've detected an optional chain, we'll be using optional accesses all the way. In
the end we are prefixing the type with `$NonMaybeType`.
* The hard-coded change only targets object types. It would be possible to target other
types but in GraphQL files these are by far the most common ones.
* I didn't put much effort to optimize this process. For every GraphQL type we're going to
pull the definition AST from shared mem and visit it. This is overall a cheap process: even
a few thousands iterations of this step shouldn't slow down the codemods beyond some
constant factor.

This approach is not specific to GraphQL, through this is by far the most common usecase.
The only real requirement is that the toplevel type is monomorphic. I would consider
extending this to any kind of file.

Changelog: [feature] codemods now serialize parts of types in `.graphql.js` files as projections of exported types from that file.

Reviewed By: SamChou19815

Differential Revision: D42421897

fbshipit-source-id: a7a77c3c59ecdc40545d73ff3080749feca3de9e