Add support for multiple hint candidates
Summary:
In some special situations of hint decomposition, we have some logic to prioritize some hints over others. e.g. for expressions like `v ?? []`, we will prioritized the hint on this entire expression for `[]` over the hint from `v`. However, this is done syntactically and we only remember the hint we have chosen, so if the hint on `v ?? []` is useless (e.g. `new Set(v ?? [])`) we will just fail and cannot recover at all.

Fundamentally, this happens because we are limited by the architecture that will only follow through a single chain of decomposition operations. We have no way to express the idea that: in case strategy A of generating hint fails, let's try strategy B. In this diff, such ability is exactly the one I will add.

I have refactored the hint representation so that the hints are now a list of hint decomposition strategies, where `[A; B; C]` means that we will try strategy `A` first, then `B`, then `C`. Only if all strategies fail, we will report failure. As a result, `Hint_None` is removed, and the same idea is now represented with an empty list, merging two hints is no longer taking the hint that is not `Hint_None`, but instead we will concat the strategy list.

Most of the changes in this diff are very mechanical, and the only interesting changes are the evaluation of hint lists and the replacement of `Hint_api.merge_hint` with `Base.List.append`

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision:
D42463613 (https://github.com/facebook/flow/commit/97042d1a060b029bba723e3d32d0b4356ec57a42)

------------------------------------------------------------------------
(from 7b6cf9d0d1219db32aa8b09e3be0796b6425a547)

fbshipit-source-id: 23f82a098370bea884effe01f5304847c5d3025d