[lti] Do not make empty arrays silently `Array<any>`
Summary:
Currently, we have a pretty serious type hole under LTI, when empty arrays can silently have type `Array<any>` if the hint on the array fails to evaluate or if the hint cannot constrain the element type. In this diff, I will fix the problem with the following strategy:

1. When hint fails to evaluate, we will emit `missing-empty-array-annot` error. This usually happens when the final hint will contain placeholders, e.g. `new Set([])`.
2. When hint evaluates to something successfully, but it still fails to constrain the element type, we will default to empty. This usually happens when the type of the array is not important (e.g. `([]: mixed)`), or if there will already be errors afterwards (e.g. tuple arity error with `([]: [number])`).

I have tried all the combinations of defaulting to empty and erroring, and found the above one to be the least disruptive and make the most sense.

Changelog: [errors] When `inference_mode=lti` is enabled, more errors related to empty arrays might be emitted. Depending on the context, Flow might either complain that the empty array needs annotation, or set the type of the empty array to be `Array<empty>`, which might cause downstream errors.

Reviewed By: panagosg7

Differential Revision: D42457376

fbshipit-source-id: ea6817da0daa34e7fb94efe8878b042111491263