More precisely track whether members are from primitive proto
Summary:
The docs for `from_proto` claimed:
> [from_proto] indicates that the member is from a primitive prototype

But this was not the case. Instead, it roughly indicated whether the prop was on the first node of the prototype chain. This meant there was no way to distinguish between being from a primitive prototype or, for example, from an inherited class.

This diff replaces `from_proto` with two new fields:
* `inherited` indicates whether the member was inherited either from a super class, higher in the prototype chain, or from an interface.
* `source` indicates whether the prop came from a primitive prototype (and which one), an interface, or just an object. This will allow us to exactly filter out these props in the next diff.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D42146101

fbshipit-source-id: 5d77365d4f3acdd6548388011c603c43060cd72c