infer empty array as Array<empty> when there is some unusable hint
Summary:
Before in cases like
```
const x: mixed { f: [] };
```
we would error asking for an annotation for `[]`. This behavior is a bit too strict and caused
a lot of new errors in codebases under LTI.

With this change the type we infer for the array in the above example is `Array<empty>`.

This diff makes more explicit the result of the reasoning around `[]` in statemnet.ml

Changelog: [internal]

(This is subsumed in the changelog message of D42457376 (https://github.com/facebook/flow/commit/786e397f98f70894129f9ea698f6ec518edb62a9).)

Reviewed By: SamChou19815

Differential Revision: D42513139

fbshipit-source-id: 7db008c7fc925c60a90f90aeec53c76626a5091b