[ts-syntax] Parse and error on `as`, `as const`, and `satisfies` cast expressions
Summary:
Changelog: [parser] Parse and error on `as`, `as const`, and `satisfies` cast expressions, to help people familiar with TS syntax.

I'm open to better ideas on implementing the parsing. These operators have the same precedence `in`/`instanceof` etc. which places them in the middle of our binary operators precedence wise - however they are not binary operators, their RHS is a type, not a value. This makes inserting their parsing a bit awkward given the existing approach.

Reviewed By: SamChou19815

Differential Revision: D42426687

fbshipit-source-id: fa1aab02e0a0fbb97a95639c992a94983b1bb3f4