[parser] fix "for (let in ...)"
Summary:
valid:
```
for (let in z) {}
```

invalid:
```
'use strict'; for (let in z) {}
```

before, we previously errored on `in` -- we try parsing it as a `for` loop with a `let` declaration with an identifier `in` and unexpected `z`, etc. this completely fails to recover.

now both examples are a `for-in` with a `let` variable. the former is valid, the latter errors about `let` being a reserved word in strict mode.

Changelog: [parser] Fix parsing of `for (let in x)` in non-strict mode: it no longer incorrectly errors.

Reviewed By: SamChou19815

Differential Revision: D42597024

fbshipit-source-id: 79b7a2bf1c831caba3838c95177ab46702baefc4